+++ R/permshap.R
@@ -16,7 +16,8 @@
 #' according to the permutation (i.e., marginalizing them over the background data).
 #' When all components are turned off, the algorithm - one by one - turns the components
 #' back on, until all components are turned on again. This antithetic scheme allows to
-#' evaluate Shapley's formula 2p times with each permutation.
+#' evaluate Shapley's formula 2p times with each permutation, using a total of
+#' 2p + 1 evaluations of marginal means.
 #'
 #' For models with interactions up to order two, one can show that
 #' even a single iteration provides exact SHAP values (with respect to the
@@ -24,10 +25,12 @@
 #'
 #' The Python implementation "shap" uses a similar approach, but without
 #' providing standard errors, and without early stopping. To mimic its behavior,
-#' we would need to set `max_iter = p` in R, and `max_eval = 2p^2` in Python.
+#' we would need to set `max_iter = p` in R, and `max_eval = (2*p+1)*p` in Python.
 #'
 #' For faster convergence, we use balanced permutations in the sense that
 #' p subsequent permutations each start with a different feature.
+#' Furthermore, the 2p on-off vectors with sum <=1 or >=p-1 are evaluated only once,
+#' similar to the degree 1 hybrid in [kernelshap()] (but covering less weight).
 #'
 #' @param exact If `TRUE`, the algorithm will produce exact SHAP values
 #'   with respect to the background data.
@@ -44,16 +47,17 @@
 #'     background data.
 #'   - `bg_X`: The background data.
 #'   - `bg_w`: The background case weights.
-#'   - `m`: Number of on-off vectors evaluated (if sampling needed: per iteration).
+#'   - `m_exact`: Number of on-off vectors evaluated once per row of `X`.
 #'   - `exact`: Logical flag indicating whether calculations are exact or not.
 #'   - `txt`: Summary text.
 #'   - `predictions`: \eqn{(n \times K)} matrix with predictions of `X`.
 #'   - `algorithm`: "permshap".
-#'   - `SE`: Standard errors corresponding to `S` (if sampling needed).
+#'   - `m`: Number of sampled on-off vectors evaluated per iteration (if not exact).
+#'   - `SE`: Standard errors corresponding to `S` (if not exact).
 #'   - `n_iter`: Integer vector of length n providing the number of iterations
-#'     per row of `X` (if sampling needed).
+#'     per row of `X` (if not exact).
 #'   - `converged`: Logical vector of length n indicating convergence per row of `X`
-#'     (if sampling needed).
+#'     (if not exact).
 #' @references
 #'   1. Erik Strumbelj and Igor Kononenko. Explaining prediction models and individual
 #'     predictions with feature contributions. Knowledge and Information Systems 41, 2014.
@@ -142,27 +146,27 @@ permshap.default <- function(
   # Pre-calculations that are identical for each row to be explained
   if (exact) {
     Z <- exact_Z(p, feature_names = feature_names, keep_extremes = TRUE)
-    m <- nrow(Z) - 2L # We won't evaluate vz for first and last row
-    m_eval <- m # for consistency with non-exact case
+    m_exact <- nrow(Z) - 2L # We won't evaluate vz for first and last row
+    m_eval <- 0L # for consistency with sampling case
     precalc <- list(
       Z = Z,
       Z_code = rowpaste(Z),
-      bg_X_rep = rep_rows(bg_X, rep.int(seq_len(bg_n), m))
+      bg_X_rep = rep_rows(bg_X, rep.int(seq_len(bg_n), m_exact))
     )
   } else {
     max_iter <- as.integer(ceiling(max_iter / p) * p) # should be multiple of p
-    m <- 2L * (p - 3L)
-    # Number of on-off vectors evaluated together in the outer loop
-    m_eval <- if (low_memory) m else m * p
+    m_exact <- 2L * p
+    m <- 2L * (p - 3L) # inner loop
+    m_eval <- if (low_memory) m else m * p # outer loop
     precalc <- list(
       bg_X_rep = rep_rows(bg_X, rep.int(seq_len(bg_n), m_eval)),
-      bg_X_balanced = rep_rows(bg_X, rep.int(seq_len(bg_n), 2L * p)),
+      bg_X_balanced = rep_rows(bg_X, rep.int(seq_len(bg_n), m_exact)),
       Z_balanced = exact_Z_balanced(p, feature_names)
     )
   }
 
-  if (m_eval * bg_n > 2e5) {
-    warning_burden(m_eval, bg_n = bg_n)
+  if (max(m_eval, m_exact) * bg_n > 2e5) {
+    warning_burden(max(m_eval, m_exact), bg_n = bg_n)
   }
 
   # Apply permutation SHAP to each row of X
@@ -217,13 +221,14 @@ permshap.default <- function(
     baseline = as.vector(v0),
     bg_X = bg_X,
     bg_w = bg_w,
-    m = m,
+    m_exact = m_exact,
     exact = exact,
     txt = txt,
     predictions = v1,
     algorithm = "permshap"
   )
   if (!exact) {
+    out$m <- m
     out$SE <- reorganize_list(lapply(res, `[[`, "sigma"))
     out$n_iter <- vapply(res, `[[`, "n_iter", FUN.VALUE = integer(1L))
     out$converged <- vapply(res, `[[`, "converged", FUN.VALUE = logical(1L))