+++ NEWS.md
@@ -12,7 +12,8 @@ from the observation to be explained), then gradually turning off components
 according to the permutation (i.e., marginalizing them over the background data).
 When all components are turned off, the algorithm - one by one - turns the components
 back on, until all components are turned on again. This antithetic scheme allows to
-evaluate Shapley's formula 2p times with each permutation.
+evaluate Shapley's formula 2p times with each permutation, using a total of
+2p + 1 evaluations of marginal means.
 
 For models with interactions up to order two, one can show that
 even a single iteration provides exact SHAP values (with respect to the
@@ -20,10 +21,13 @@ given background dataset).
 
 The Python implementation "shap" uses a similar approach, but without
 providing standard errors, and without early stopping. To mimic its behavior,
-we would need to set `max_iter = p` in R, and `max_eval = 2p^2` in Python.
+we would need to set `max_iter = p` in R, and `max_eval = (2*p+1)*p` in Python.
 
 For faster convergence, we use balanced permutations in the sense that
 p subsequent permutations each start with a different feature.
+Furthermore, the 2p on-off vectors with sum <=1 or >=p-1 are evaluated only once,
+similar to the degree 1 hybrid in [kernelshap()] (but covering less weight).
+
 ### User visible changes
 
 - In exact mode, `kernelshap()` does not return the following elements anymore:
@@ -41,6 +45,10 @@ p subsequent permutations each start with a different feature.
 ### Maintenance
 
 - Improve code coverage ([#156](https://github.com/ModelOriented/kernelshap/pull/156)).
+
+### Bug fixes
+
+- `kernelshap()` with `max_iter = 1` will now work ([#160](https://github.com/ModelOriented/kernelshap/pull/160)).
 
 # kernelshap 0.7.1 (not on CRAN)